+++ plugins/source/aws/resources/services/wellarchitected/share_invitations_mock_test.go
@@ -3,6 +3,7 @@ package wellarchitected
 import (
 	"testing"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
@@ -21,7 +22,7 @@ func buildShareInvitationsMock(t *testing.T, ctrl *gomock.Controller) client.Ser
 
 		m.EXPECT().ListShareInvitations(gomock.Any(),
 			&wellarchitected.ListShareInvitationsInput{
-				MaxResults:        50,
+				MaxResults:        aws.Int32(50),
 				ShareResourceType: shareResourceType,
 			}, gomock.Any()).
 			Return(