+++ plugins/source/aws/resources/services/wellarchitected/workload_shares_mock_test.go
@@ -3,6 +3,7 @@ package wellarchitected
 import (
 	"testing"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client/mocks"
@@ -18,7 +19,7 @@ func buildWorkloadShares(t *testing.T, m *mocks.MockWellarchitectedClient, workl
 	m.EXPECT().ListWorkloadShares(gomock.Any(),
 		&wellarchitected.ListWorkloadSharesInput{
 			WorkloadId: workload.WorkloadId,
-			MaxResults: 50,
+			MaxResults: aws.Int32(50),
 		},
 		gomock.Any()).
 		Return(