+++ plugins/source/aws/resources/services/wellarchitected/workloads_mock_test.go
@@ -3,6 +3,7 @@ package wellarchitected
 import (
 	"testing"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected"
 	"github.com/aws/aws-sdk-go-v2/service/wellarchitected/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
@@ -19,7 +20,7 @@ func buildWorkloadsMock(t *testing.T, ctrl *gomock.Controller) client.Services {
 	require.NoError(t, faker.FakeObject(&summary))
 
 	m.EXPECT().ListWorkloads(gomock.Any(),
-		&wellarchitected.ListWorkloadsInput{MaxResults: 50},
+		&wellarchitected.ListWorkloadsInput{MaxResults: aws.Int32(50)},
 		gomock.Any()).
 		Return(&wellarchitected.ListWorkloadsOutput{WorkloadSummaries: []types.WorkloadSummary{summary}}, nil)